Fix colortab

+++ go-client/colortab/colortab.go
@@ -22,9 +22,9 @@ func main() {
 		outline            = flag.Bool("outline", false, "outline swatch")
 		fontsize           = flag.Int("fs", 12, "fontsize")
 		rowsize            = flag.Int("r", 32, "rowsize")
-		colw               = flag.Float32("c", 340, "column size")
-		swatch             = flag.Float32("s", 16, "swatch size")
-		gutter             = flag.Float32("g", 12, "gutter")
+		colw               = flag.Float64("c", 340, "column size")
+		swatch             = flag.Float64("s", 16, "swatch size")
+		gutter             = flag.Float64("g", 12, "gutter")
 		err          error
 		tcolor, line string
 	)
@@ -49,15 +49,15 @@ func main() {
 		openvg.Rect(0, 0, fw, fh)
 		tcolor = "black"
 	}
-	top := fh - 32
-	left := 32.0
+	top := fh - 32.0
+	left := float32(32.0)
 	in := bufio.NewReader(f)
 
 	for x, y, nr := left, top, 0; err == nil; nr++ {
 		line, err = in.ReadString('\n')
 		fields := strings.Split(strings.TrimSpace(line), "\t")
 		if nr%*rowsize == 0 && nr > 0 {
-			x += *colw
+			x += float32(*colw)
 			y = top
 		}
 		if len(fields) == 3 {
@@ -69,13 +69,13 @@ func main() {
 				openvg.StrokeWidth(1)
 			}
 			if *circsw {
-				openvg.Circle(x+*swatch/2, y+*swatch/2, *swatch)
+				openvg.Circle(x+float32(*swatch)/2.0, y+float32(*swatch)/2.0, float32(*swatch))
 			} else {
-				openvg.Rect(x, y, *swatch, *swatch)
+				openvg.Rect(x, y, float32(*swatch), float32(*swatch))
 			}
 			openvg.StrokeWidth(0)
 			openvg.FillColor(tcolor)
-			openvg.Text(x+*swatch+float32(*fontsize/2), y, fields[0], *fontname, *fontsize)
+			openvg.Text(x+float32(*swatch)+float32(*fontsize/2), y, fields[0], *fontname, *fontsize)
 			var label string
 			if *showcode {
 				if *showrgb {
@@ -84,10 +84,10 @@ func main() {
 					label = fields[2]
 				}
 				openvg.FillColor("gray")
-				openvg.TextEnd(x+*colw-(*swatch+*gutter), y, label, *fontname, *fontsize)
+				openvg.TextEnd(x+float32(*colw)-(float32(*swatch)+float32(*gutter)), y, label, *fontname, *fontsize)
 			}
 		}
-		y -= (*swatch + *gutter)
+		y -= float32(*swatch) + float32(*gutter)
 	}
 	openvg.End()
 	bufio.NewReader(os.Stdin).ReadBytes('\n')
